functions scope & cleaning
