+++ src/5gmsaf/metrics-reporting-provisioning.h
@@ -1,4 +1,3 @@
-
 #ifndef MSAF_METRICS_REPORTING_PROVISIONING_H
 #define MSAF_METRICS_REPORTING_PROVISIONING_H
 
@@ -16,6 +15,10 @@ typedef struct msaf_metrics_reporting_configuration_s {
     OpenAPI_list_t *metrics;
 } msaf_metrics_reporting_configuration_t;
 
-extern OpenAPI_metrics_reporting_configuration_t *msaf_metrics_reporting_configuration();
+// Functions declarations
+msaf_metrics_reporting_configuration_t *msaf_metrics_reporting_configuration_create();
+void msaf_delete_metrics_reporting_configuration(const char *provisioning_session_id);
+msaf_metrics_reporting_configuration_t *msaf_metrics_reporting_configuration_find_by_metricsReportingConfigurationId(const char *metricsReportingConfigurationId);
+cJSON *msaf_get_metrics_reporting_configuration_by_provisioning_session_id(const char *provisioning_session_id);
 
 #endif //MSAF_METRICS_REPORTING_PROVISIONING_H
